Adapt to highlighting definition schema changing version to an integer
We only still have to consider floats when loading legacy files from Kate,
those will now implicitly be rounded down to the next integer version.

Fixes T4050.

+++ src/indexer/katehighlightingindexer.cpp
@@ -129,7 +129,7 @@ int main(int argc, char *argv[])
         }
 
         // numerical attributes
-        hl[QStringLiteral("version")] = xml.attributes().value(QLatin1String("version")).toFloat();
+        hl[QStringLiteral("version")] = xml.attributes().value(QLatin1String("version")).toInt();
         hl[QStringLiteral("priority")] = xml.attributes().value(QLatin1String("priority")).toInt();
 
         // add boolean one
